Move WemoData/WemoConfigEntryData to models.py
